# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Blog Personnel - ISEP_P5",
        "description": "Cette API permet la gestion d'un blog personnel : authentification par OTP, gestion des articles, des amis et des commentaires.",
        "contact": {
            "name": "Équipe ISEP",
            "email": "support@isep.sn"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local (dev)"
        },
        {
            "url": "https://api.monblog.isep.sn",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/users/search": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Rechercher des utilisateurs",
                "operationId": "a25d297a6c0498e9254b9dde5a8a9e7d",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs trouvés"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/friends/request": {
            "post": {
                "tags": [
                    "Amis"
                ],
                "summary": "Envoyer une demande d'amitié",
                "operationId": "cc068cd68371d2e80cdc1f163b91390b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "friend_id"
                                ],
                                "properties": {
                                    "friend_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Demande d'amitié envoyée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/friends": {
            "get": {
                "tags": [
                    "Amis"
                ],
                "summary": "Récupérer la liste des amis",
                "operationId": "de299b3adac147e2534574bc1cb0db60",
                "responses": {
                    "200": {
                        "description": "Liste des amis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/friends/accept": {
            "post": {
                "tags": [
                    "Amis"
                ],
                "summary": "Accepter une demande d'amitié",
                "operationId": "4ec5ca4d7b6755971039fa45d4672a41",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "friend_id"
                                ],
                                "properties": {
                                    "friend_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Demande d'amitié acceptée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/friends/reject": {
            "post": {
                "tags": [
                    "Amis"
                ],
                "summary": "Rejeter une demande d'amitié",
                "operationId": "61268f0b4d6886a06e4aeee82c8f1482",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "friend_id"
                                ],
                                "properties": {
                                    "friend_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Demande d'amitié rejetée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/friends/{friend_id}": {
            "delete": {
                "tags": [
                    "Amis"
                ],
                "summary": "Supprimer un ami",
                "operationId": "a62435b0c0f400d9a4210ccb222193d1",
                "parameters": [
                    {
                        "name": "friend_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ami supprimé avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/friends/block": {
            "post": {
                "tags": [
                    "Amis"
                ],
                "summary": "Bloquer un utilisateur",
                "operationId": "8a581737fc2af03f9f755f599ce068e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "friend_id"
                                ],
                                "properties": {
                                    "friend_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur bloqué avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}